src/requirements.txt
requests
SQLAlchemy
psycopg2-binary
aiohttp
coverage
testing.postgresql